requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --output-file requirements.txt --no-dev --no-hashes
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
